requirements.txt
alembic==1.14.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
Flask==2.2.2
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-MySQL==1.6.0
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
PyJWT==2.9.0
PyMySQL==1.1.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==2.2.2
zipp==3.21.0
